Split Register Form to separate client component from page
