+++ app/auth/components/form.tsx
@@ -9,9 +9,9 @@ import { Button } from "@/components/ui/button"
 import { useSearchParams } from "next/navigation"
 import { toast } from "@/components/ui/use-toast"
 import { useFormStatus } from "react-dom"
-import { login } from "@/app/api/auth/action"
+import { login, signup } from "@/app/api/auth/action"
 
-const formSchema = z.object({
+const loginFormSchema = z.object({
     email: z.string(),
     password: z.string(),
 })
@@ -41,8 +41,8 @@ export function LoginForm() {
 
     }, [searchParams]);
 
-    const form = useForm<z.infer<typeof formSchema>>({
-        resolver: zodResolver(formSchema),
+    const form = useForm<z.infer<typeof loginFormSchema>>({
+        resolver: zodResolver(loginFormSchema),
         defaultValues: {
             email: "",
             password: "",
@@ -107,6 +107,165 @@ export function LoginForm() {
                         </div>
                         <div className="grid gap-2">
                             <SubmitButton />
+                        </div>
+
+                    </div>
+                </form>
+            </Form>
+        </div>
+    )
+}
+
+const registerFormSchema = z.object({
+    fullName: z.string(),
+    email: z.string(),
+    password: z.string(),
+    passwordConfirm: z.string(),
+})
+
+export function RegisterForm() {
+    const [passwordVisibility, setPasswordVisibility] = useState([false, false]);
+    const searchParams = useSearchParams()
+
+    const form = useForm<z.infer<typeof registerFormSchema>>({
+        resolver: zodResolver(registerFormSchema),
+        defaultValues: {
+            fullName: "",
+            email: "",
+            password: "",
+            passwordConfirm: "",
+        },
+    })
+
+    useEffect(()=> {
+        if(searchParams.get('error')){
+            toast({
+                title: "Invalid Credential",
+                duration: 2000
+            })
+        }
+    }, [searchParams])
+
+    const submitLogin = useCallback(async (formData: FormData) => {
+        await signup(formData);
+
+        if(searchParams.get('error')){
+            toast({
+                title: "Invalid Credential",
+                duration: 2000
+            })
+        }
+
+    }, [searchParams]);
+
+    return (
+        <div className="py-4">
+            <Form  {...form}>
+                <form  action={submitLogin} className="mx-4 p-4 lg:mx-auto max-w-3xl grid gap-4 border rounded">
+                    <div className="grid gap-2">
+                        <div className="py-4">
+                            <h2 className="block text-2xl text-center font-medium">Register Form</h2>
+                        </div>
+                        <div className="grid gap-2">
+                            <FormField
+                                control={form.control}
+                                name="fullName"
+                                render={({ field }) => (
+                                    <FormItem>
+                                        <FormLabel>Full Name</FormLabel>
+                                        <FormControl>
+                                            <Input placeholder="" {...field} />
+                                        </FormControl>
+                                        <FormDescription />
+                                        <FormMessage />
+                                    </FormItem>
+                                )}
+                            />
+                        </div>
+
+                        <div className="grid gap-2">
+                            <FormField
+                                control={form.control}
+                                name="email"
+                                render={({ field }) => (
+                                    <FormItem>
+                                        <FormLabel>Email</FormLabel>
+                                        <FormControl>
+                                            <Input type="email" placeholder="" {...field} />
+                                        </FormControl>
+                                        <FormDescription />
+                                        <FormMessage />
+                                    </FormItem>
+                                )}
+                            />
+                        </div>
+
+                        <div className="grid grid-cols-12 gap-2">
+                            <FormField
+                                control={form.control}
+                                name="password"
+                                render={({ field }) => (
+                                    <FormItem className="col-span-11">
+                                        <FormLabel>Password</FormLabel>
+                                        <FormControl>
+                                            <Input type={passwordVisibility[0] ? "text" : "password"} placeholder="" {...field} />
+                                        </FormControl>
+                                        <FormDescription />
+                                        <FormMessage />
+                                    </FormItem>
+                                )}
+                            />
+                            <Button onClick={(e)=> {e.preventDefault(); setPasswordVisibility([!passwordVisibility[0], passwordVisibility[1]])}} variant={"outline"} className="self-end mb-2">
+                                {
+                                    passwordVisibility[0] ? (
+                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="size-6">
+                                            <path strokeLinecap="round" strokeLinejoin="round" d="M3.98 8.223A10.477 10.477 0 0 0 1.934 12C3.226 16.338 7.244 19.5 12 19.5c.993 0 1.953-.138 2.863-.395M6.228 6.228A10.451 10.451 0 0 1 12 4.5c4.756 0 8.773 3.162 10.065 7.498a10.522 10.522 0 0 1-4.293 5.774M6.228 6.228 3 3m3.228 3.228 3.65 3.65m7.894 7.894L21 21m-3.228-3.228-3.65-3.65m0 0a3 3 0 1 0-4.243-4.243m4.242 4.242L9.88 9.88" />
+                                        </svg>
+
+                                    ) : (
+                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="size-6">
+                                            <path strokeLinecap="round" strokeLinejoin="round" d="M2.036 12.322a1.012 1.012 0 0 1 0-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178Z" />
+                                            <path strokeLinecap="round" strokeLinejoin="round" d="M15 12a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z" />
+                                        </svg>
+                                    )
+                                }    
+                            </Button>
+                        </div>
+                        
+                        <div className="grid grid-cols-12 gap-2">
+                            <FormField
+                                control={form.control}
+                                name="passwordConfirm"
+                                render={({ field }) => (
+                                    <FormItem className="col-span-11">
+                                        <FormLabel>Password Confirmation</FormLabel>
+                                        <FormControl>
+                                            <Input type={passwordVisibility[1] ? "text" : "password"} placeholder="" {...field} />
+                                        </FormControl>
+                                        <FormDescription />
+                                        <FormMessage />
+                                    </FormItem>
+                                )}
+                            />
+                            <Button onClick={(e)=> {e.preventDefault(); setPasswordVisibility([passwordVisibility[0], !passwordVisibility[1]])}} variant={"outline"} className="self-end mb-2">
+                                {
+                                    passwordVisibility[1] ? (
+                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="size-6">
+                                            <path strokeLinecap="round" strokeLinejoin="round" d="M3.98 8.223A10.477 10.477 0 0 0 1.934 12C3.226 16.338 7.244 19.5 12 19.5c.993 0 1.953-.138 2.863-.395M6.228 6.228A10.451 10.451 0 0 1 12 4.5c4.756 0 8.773 3.162 10.065 7.498a10.522 10.522 0 0 1-4.293 5.774M6.228 6.228 3 3m3.228 3.228 3.65 3.65m7.894 7.894L21 21m-3.228-3.228-3.65-3.65m0 0a3 3 0 1 0-4.243-4.243m4.242 4.242L9.88 9.88" />
+                                        </svg>
+
+                                    ) : (
+                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="size-6">
+                                            <path strokeLinecap="round" strokeLinejoin="round" d="M2.036 12.322a1.012 1.012 0 0 1 0-.639C3.423 7.51 7.36 4.5 12 4.5c4.638 0 8.573 3.007 9.963 7.178.07.207.07.431 0 .639C20.577 16.49 16.64 19.5 12 19.5c-4.638 0-8.573-3.007-9.963-7.178Z" />
+                                            <path strokeLinecap="round" strokeLinejoin="round" d="M15 12a3 3 0 1 1-6 0 3 3 0 0 1 6 0Z" />
+                                        </svg>
+                                    )
+                                }    
+                            </Button>
+                        </div>
+                        
+                        <div className="grid gap-2">
+                            <Button type="submit">Submit</Button>
                         </div>
 
                     </div>
